js: modularize some more, check in main.js if journal needs to be initialized
