+++ fava/templates/context.html
@@ -5,7 +5,6 @@
 
 {% block javascript %}
     {% include "javascript/_editor.html" %}
-    {% include "javascript/_journal.html" %}
 {% endblock %}
 
 {% block content %}
