+++ fava/templates/documents.html
@@ -1,8 +1,6 @@
 {% extends "_layout.html" %}
 {% set active_page = 'documents' %}
 
-{% block javascript %}{% include "javascript/_journal.html" %}{% endblock %}
-
 {% block content %}
     <h1>Documents</h1>
     {% set documents = api.documents() %}
