--- fava/templates/javascript/_journal.html
@@ -1,1 +1,0 @@
-<script type="text/javascript" src="{{ url_for('static', filename='gen/journal.js') }}"></script>
