+++ fava/templates/journal.html
@@ -1,8 +1,6 @@
 {% extends "_layout.html" %}
 {% set active_page = 'journal' %}
 
-{% block javascript %}{% include "javascript/_journal.html" %}{% endblock %}
-
 {% block content %}
     <h1>General Ledger</h1>
     {% with journal = api.journal(), show_tablefilter=True, show_change_and_balance=config.user.getboolean('journal-general-show-balances')  %}
